% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cn.R
\name{font_cn}
\alias{font_cn}
\title{cn font for ggplot2}
\usage{
font_cn(font = "STKaiti")
}
\description{
cn font for ggplot2
}
\examples{
library(ggformula)
gf_point(mpg ~ hp, color = ~ cyl, size = ~wt, data = mtcars) -> p
p
p + xlab("X轴") + ylab("Y轴") + ggtitle("标题") # + font_cn()
p + xlab("X轴") + ylab("Y轴") + ggtitle("标题") + font_cn()
}
