% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_cn.R
\name{theme_cn}
\alias{theme_cn}
\title{cn font for ggplot2}
\usage{
theme_cn(font = "STKaiti")
}
\description{
cn font for ggplot2
}
\examples{
library(ggformula)
gf_point(mpg ~ hp, color = ~ cyl, size = ~wt, data = mtcars) -> p
p
p + xlab("X轴") + ylab("Y轴") + ggtitle("标题") # + theme_cn()
p + xlab("X轴") + ylab("Y轴") + ggtitle("标题") + theme_cn()
}
